% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_bxcx.R
\name{apply_bxcx}
\alias{apply_bxcx}
\title{Apply Box-Cox transformation to rval
  Zachary Rowson
  Rowson.Zachary at epa.gov
  Last edit: 05/14/2021}
\usage{
apply_bxcx(data)
}
\arguments{
\item{data}{is a mc0 dataset formatted as below.
\itemize{
  \item srcf - name of file that is being formatted
  \item acid - assay component id (Here ZFpmrA/L/D-20-40-40)
    Zebrafish photomotor resonse, Accilmation/Light/Dark-20 minutes-40 minutes-40 minutes
  \item cpid - chemical name
  \item apid - assay plate id DNT###
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
   t = test, v = vehicle control
  \item wllq - well quality indicates if observation is viable for analysis
    1 = yes, 0 = no
  \item conc - concentration of chemical
  \item rval - endpoint of interest
}}
}
\value{
mc0 formatted table with Box-Cox tranformed rvals

  @import data.table
}
\description{
Applies Box-Cox power transformation to mc0 rval with goal
of normalizing data. Requires estimating an optimal
lambda paramater, lambda.hat. Lambda.hat is chosen via
MLE estimation using control data.
}
