% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mc0.R
\name{as_mc0}
\alias{as_mc0}
\title{This function reformats behavioral data as mc0. Much of this function copies code from
Katie Paul-Friedmans "PADILLA_ZF_BEHAVIOR_s_to_tcpl(editted).R"
Future updates to gabi will likey not require use
of this function.}
\usage{
as_mc0(chemnames, endp)
}
\arguments{
\item{chemnames}{vector of chemical names as strings}

\item{endp}{string name of endpoint to saved as rval}
}
\description{
This function reformats behavioral data as mc0. Much of this function copies code from
Katie Paul-Friedmans "PADILLA_ZF_BEHAVIOR_s_to_tcpl(editted).R"
Future updates to gabi will likey not require use
of this function.
}
