% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pmr0.R
\name{as_pmr0}
\alias{as_pmr0}
\title{Format data as pmr0
  Zachary Rowson
  Rowson.Zachary at epa.gov
  Last edit: 04/06/2021}
\usage{
as_pmr0(data)
}
\arguments{
\item{data}{is a single sheet from "All Chemicals one per sheet.xlsx" as a data.table}
}
\value{
A pmr0 format data.table
  \itemize{
    \item srcf - name of file that is being formatted
    \item acid - assay component id (Here ZFpmrA/L/D-20-40-40)
      Zebrafish photomotor resonse, Accilmation/Light/Dark-20 minutes-40 minutes-40 minutes
    \item cpid - chemical name
    \item apid - assay plate id DNT###
    \item rowi - row on plate
    \item coli - column on plate
    \item wllt - well type according to tcpl mc0 format
     t = test, v = vehicle control
    \item wllq - well quality indicates if observation is viable for analysis
      1 = yes, 0 = no
    \item conc - concentration of chemical
    \item vtj - measurements at time period j for j /in {1,2,...,n}
      Here n = 50
  }

  @import data.table
}
\description{
Formats a US EPA ORD CCTE BCTD RADB Padilla lab DNT60 data
set into pmr0 format. pmr0 format is easily transformed
into mc0 with gabi::as_mc0() for tcpl analysis.
}
