% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_row.R
\name{as_row}
\alias{as_row}
\title{This function produces a list in format of tcplfit2 "row" for named chemicals
  from a data.table formatted as mc0.}
\usage{
as_row(table, chemicals, ctoff, ratio = FALSE, dist = FALSE, global = TRUE)
}
\arguments{
\item{table}{is a mc0 level tcpl dataset.}

\item{chemicals}{is a vector of strings indicating test chemical data to be processed.}

\item{ctoff}{is a decimal in (0,1) or a string indicating the statistic to be used as a cutoff value.}

\item{ratio}{is a logical parameter indicating return of resp in log(test/mean(baseline)) form.}

\item{dist}{is a logical parameter indicating return of resp in resp-mean(baseline) format.}

\item{global}{is a logical parameter with TRUE indicating assignment in gloabl environment.}
}
\description{
This function produces a list in format of tcplfit2 "row" for named chemicals
  from a data.table formatted as mc0.
}
