% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AUC.R
\name{calc_AUC}
\alias{calc_AUC}
\title{Calculate AUC in light and dark for individual fish}
\usage{
calc_AUC(data, no.A = 10, no.L = 20, no.D = 20)
}
\arguments{
\item{data}{is a pmr0 formatted data.table
\itemize{
  \item srcf - name of file that is being formatted
  \item acid - assay component id
  \item cpid - chemical name
  \item apid - assay plate id
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
  \item wllq - well quality indicates if observation is viable for analysis
  \item conc - concentration of chemical
  \item tj - measurements at time period j for j /in {1,2,3,...,n}
}}

\item{no.A}{number of measurements made in acclimation period}

\item{no.L}{number of measurements made in light period}

\item{no.D}{number of measurements made in dark period}
}
\value{
A list with AUC in each light level for each fish
  \itemize{
    \item AUC_L - AUC in light
    \item AUC_D - AUC in dark
    \item AUC_r - AUC_D / AUC_L
 }
}
\description{
Calculates area under curve for a chosen light level (light or dark).
Uses trapezoidal rule to approximate integral of speed function.
User provides a pmr0 table and specifies number of time periods per
experimental stage: acclimation, light, and dark.
}
\details{
Last edit 08/18/2022.
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
