% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_avgS.R
\name{calc_avgS}
\alias{calc_avgS}
\title{Calculate average speed per light level for individual fish}
\usage{
calc_avgS(data, no.A = 10, no.L = 20, no.D = 20)
}
\arguments{
\item{data}{is a pmr0 formatted data.table
\itemize{
  \item srcf - name of file that is being formatted
  \item acid - assay component id
  \item cpid - chemical name
  \item apid - assay plate id
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
  \item wllq - well quality indicates if observation is viable for analysis
  \item conc - concentration of chemical
  \item tj - measurements at time period j for j /in {1,2,3,...,n}
}}

\item{no.A}{number of measurements made in acclimation period}

\item{no.L}{number of measurements made in light period}

\item{no.D}{number of measurements made in dark period}
}
\value{
A list with average speed in each light level for each fish
  \itemize{
    \item avgS_L - average speed in light
    \item avgS_D - average speed in dark
  }
}
\description{
Calculates average speed per light level (dark or light) for
individual fish. User provides a lmr0 table and inputs
number of time periods for each experimental stage:
acclimation, light, and dark. A lmr0 table
}
\details{
Last edit 08/18/2022.
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
