% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_endp.R
\name{calc_endp}
\alias{calc_endp}
\title{Calculate endpoint values from lmr0 formatted table}
\usage{
calc_endp(data, no.A = 10, no.L = 20, no.D = 20, rval = NULL)
}
\arguments{
\item{data}{is a lmr0 formatted data.table
\itemize{
  \item srcf - name of file that is being formatted
  \item acid - assay component id
  \item cpid - chemical name
  \item apid - assay plate id
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
  \item wllq - well quality indicates if observation is viable for analysis
  \item conc - concentration of chemical
  \item tj - measurements at time period j for j /in {1,2,3,...,n}
}}

\item{no.A}{number of measurements made in acclimation period}

\item{no.L}{number of measurements made in light period}

\item{no.D}{number of measurements made in dark period}

\item{rval}{- string representing endpoint of interest. Can equal...
\itemize{
  \item AUC_L - area under curve in light
  \item AUC_D - area under curve in dark
  \item AUC_T - area under curve over whole time-series
  \item AUC_r - ratio of AUC_D/AUC_L
  \item avgS_L - average speed in light
  \item avgS_D - average speed in dark
  \item avgS_T - average speed over whole time series
  \item avgA_L - average acceleration in light
  \item avgA_D - average acceleration in dark
  \item avgJ_L - average jerk in light
  \item avgJ_D - average jerk in dark
  \item strtlA - acceleration during startle, light-dark, transition
  \item strtlAavg - difference between apeed in last period of dark and avgS_L
  \item strtlF - startle factor, ratio of speed in first period of dark over speed in last period of light
  \item frzA - acceleration during freeze, dark-light, transition
  \item frzF - freeze factor, ratio of speed in first period of light over speed in last period of acclimation
}}
}
\value{
A vector with user specified endpoint values
}
\description{
Calculates endpoint values for tcpl analysis from pmr0
formatted table. Function is a wrapper function for
endpoint calculating functions in calc_ family. Endpoint
values fill rval column in mc0 table after
conversion with gabi::as_mc0.
}
\details{
Details on what endpoints represent can be found
by searching for appropriate calc_ function (?gabi::calc_...).
Last edit 08/18/2022.
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
