% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hbt.R
\name{calc_hbt}
\alias{calc_hbt}
\title{Calculate Habituation Measurements}
\usage{
calc_hbt(data, no.A = 10, no.L = 20, no.D = 20)
}
\arguments{
\item{data}{is a lmr0 formatted data.table
\itemize{
  \item srcf - name of file that is being formatted
  \item acid - assay component id
  \item cpid - chemical name
  \item apid - assay plate id
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
  \item wllq - well quality indicates if observation is viable for analysis
  \item conc - concentration of chemical
  \item tj - measurements at time period j for j /in {1,2,3,...,n}
}}

\item{no.A}{number of measurements made in acclimation period}

\item{no.L}{number of measurements made in light period}

\item{no.D}{number of measurements made in dark period}
}
\value{
A list with average jerk in each light level for each fish
  \itemize{
    \item hbt1_L - average acceleration in light
    \item hbt1_D - average acceleration in dark
    \item hbt2_L - average jerk in light
    \item hbt2_D - average jerk in dark
  }
}
\description{
Calculates approximation slope of activity trends with time (Habituation 1)
 and approximate curvature of activity trends with time (Habituation 2)
(dark or Light) for individual fish. User provides a lmr0 table and inputs
number of time periods for each experimental stage: Acclimation,
Light, and Dark.
}
\details{
Excludes Habituation 1 and Habituation 2 values during transitions
(freeze and startle) from calculation of averages.
Last edit 11/15/2022.
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
