% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_qCI.R, R/data_qCI.R
\name{calc_qCI}
\alias{calc_qCI}
\title{Calculate confidence intervals for a quantile}
\usage{
calc_qCI(x, quantile = 50, conf.level = 0.95)

calc_qCI(x, quantile = 50, conf.level = 0.95)
}
\arguments{
\item{x}{is a vector of sample values}

\item{quantile}{is a numeric representing quantile for which a CI is
calculated. Defaults to median or 50% quantile.}

\item{conf.level}{is a numeric indicator of desired confidence-level
of constructed interval.}
}
\value{
list
  \itemize{
    \item interval - vector of upper and lower bounds of confidence interval
    \item confidence - confidence level associated with calculated interval
    \item order.stats - location of interval bounds in ordered sample
   }

list
  \itemize{
    \item interval - vector of upper and lower bounds of confidence interval
    \item confidence - confidence level associated with calculated interval
    \item order.stats - location of interval bounds in ordered sample
   }
}
\description{
Calculates confidence intervals at specified confidence-level for a
quantile of users choosing using empirical order-statistics.

Calculates confidence intervals at specified confidence-level for a
quantile of users choosing using empirical order-statistics.
}
\details{
Last edit: 06/04/2021
Roxygen created this manual page on ``r Sys.Date()`` using R version
``r getRversion()``.
Code written by Arthur Charpentier taken from his post on
freakonometrics \url{https://freakonometrics.hypotheses.org/4199}
Was checked by Zachary Rowson for correctness.

Last edit: 06/04/2021
Roxygen created this manual page on ``r Sys.Date()`` using R version
``r getRversion()``.
Code written by Arthur Charpentier taken from his post on
freakonometrics \url{https://freakonometrics.hypotheses.org/4199}
Was checked by Zachary Rowson for correctness.
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}

Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
