% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_RoA.R
\name{calc_RoA}
\alias{calc_RoA}
\title{Calculate Range of Activity per light level for individual fish}
\usage{
calc_RoA(data, no.A = 10, no.L = 20, no.D = 20)
}
\arguments{
\item{data}{is a pmr0 formatted data.table
\itemize{
  \item srcf - name of file that is being formatted
  \item acid - assay component id
  \item cpid - chemical name
  \item apid - assay plate id
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
  \item wllq - well quality indicates if observation is viable for analysis
  \item conc - concentration of chemical
  \item tj - measurements at time period j for j /in {1,2,3,...,n}
}}

\item{no.A}{number of measurements made in acclimation period}

\item{no.L}{number of measurements made in light period}

\item{no.D}{number of measurements made in dark period}
}
\value{
A list with Range of Activity endpoints for each fish
  \itemize{
    \item RoA_L - speed ratio in light. max(speed) / min(speed)
    \item RoA_D - speed ratio in dark. max(speed) / min(speed)
  }
}
\description{
Calculates range of activity measurements per light period (light or dark).
User provides a lmr0 table and inputs
number of time periods in each experimental stage:
acclimation, light, and dark.
}
\details{
RoA_L = max(speed in light) / min(speed in light), RoA_D = max(speed in dark) / min(speed in dark).
Created 09/28/2021. Last edit 11/15/2022.
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
