% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_trans.R
\name{calc_trans}
\alias{calc_trans}
\title{Calculate transition endpoints}
\usage{
calc_trans(data, no.A = 10, no.L = 20, no.D = 20)
}
\arguments{
\item{data}{is a lmr0 formatted data.table
\itemize{
  \item srcf - name of file that is being formatted
  \item acid - assay component id
  \item cpid - chemical name
  \item apid - assay plate id
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
  \item wllq - well quality indicates if observation is viable for analysis
  \item conc - concentration of chemical
  \item tj - measurements at time period j for j /in {1,2,3,...,n}
}}

\item{no.A}{number of measurements made in acclimation period}

\item{no.L}{number of measurements made in light period}

\item{no.D}{number of measurements made in dark period}
}
\value{
A list with vectors of transitory endpoint values for each fish
  \itemize{
    \item strtlA - "startle acceleration" acceleration at startle
    \item strtlAavg - "startle acceleration from average" acceleration at startle when speed in dark is compared to average speed in light
    \item strtlF - "startle factor" fold change in speed at startle
  }
}
\description{
Calculates transition endpoints from user provided behavioral data.
Transition endpoints describe fish behavior at transitions between
different light levels: startle (light to dark).
}
\details{
Last edit 08/18/2022.
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
