% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Endp.R
\name{create_Endp}
\alias{create_Endp}
\title{Create endpoints from pmr0 formatted table
  Zachary Rowson
  Rowson.Zachary at epa.gov
  Last edit: 04/06/2021}
\usage{
create_Endp(table, no.A = 10, no.L = 20, no.D = 20)
}
\arguments{
\item{table}{is a pmr0 formatted data.table
\itemize{
  \item srcf - name of file that is being formatted
  \item acid - assay component id
  \item cpid - chemical name
  \item apid - assay plate id
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
  \item wllq - well quality indicates if observation is viable for analysis
  \item conc - concentration of chemical
  \item tj - measurements at time period j for j /in {1,2,3,...,n}
}}

\item{no.A}{number of measurements made in acclimation period}

\item{no.L}{number of measurements made in light period}

\item{no.D}{number of measurements made in dark period}
}
\value{
A table with endpoint measurements
  \itemize{
    \item ...first columns same as pmr0 format minus tj data
    \item
  }
  @import data.table
}
\description{
Returns complete USA EPA ORD CCTE BCTD RADB Padilla lab DNT60 data
set in pmr0 format. pmr0 format is easily transformed
into mc0 with DNT.60.Analysis::as_mc0() for tcpl analysis.
}
