% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Hitcalls.R
\name{create_Hitcalls}
\alias{create_Hitcalls}
\title{This function uses tcplFit2 to create a table of hitcalls for chemicals tested in various assays.
Currently limited to execution on one chemical.}
\usage{
create_Hitcalls(tables, chemicals, ctoff)
}
\arguments{
\item{tables}{is a list of mc0 level tcpl datasets}

\item{chemicals}{is vector of chemical names as they appear in mc0}

\item{ctoff}{is an object describing the cutoff metric to be used. Can be string "MAD", or numerics 0.01, 0.05, 0.95, or 0.99.}
}
\description{
This function uses tcplFit2 to create a table of hitcalls for chemicals tested in various assays.
Currently limited to execution on one chemical.
}
