% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eDist.R
\name{plot_eDist}
\alias{plot_eDist}
\title{Plot Endpoint Distributions by Concentration Group as Scatter Plots}
\usage{
plot_eDist(
  mc0,
  chemical,
  unit.conc = paste0("μ", "M"),
  plot.log = TRUE,
  lam.hat = "NA",
  shift = "NA"
)
}
\arguments{
\item{mc0}{- is a mc0 dataset formatted as below
\itemize{
  \item srcf - name of file holding raw data
  \item acid - assay component id (Here ZFlmrALD-20-40-40)
    Zebrafish locomotor resonse, Acclimation/Light/Dark- _l_ minutes- _m_ minutes- _n_ minutes
  \item cpid - chemical name
  \item apid - assay plate id DNT###
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
   t = test, v = vehicle control
  \item wllq - well quality indicates if observation is viable for analysis
   1 = yes, 0 = no
  \item conc - concentration of chemical
  \item rval - endpoint values for fish
}}

\item{chemical}{- string corresponding to name of chemical of interest}

\item{unit.conc}{- unit of chemical concentration. Defaults to $\mu$M}

\item{plot.log}{- boolean indicator of whether or not concentrations should be plotted on a
logarithmic, base 10, scale. Defaults to TRUE}

\item{lam.hat}{- power parameter used in Box-Cox transformation of the data}

\item{shift}{- shift parameter used in Box-Cox transformation}
}
\value{
A ggplot2 object displaying scatterplots of endpoint values grouped by concentration
}
\description{
Plots endpoint data for a specific chemical and corresponding control
as scatter-plots grouped by chemical concentration. Takes as arguments
data in mc0 format and a chemical name.
}
\details{
Created: 10/12/2021
Last edit: 11/15/2021

Scatter-plots are provided rather than box-plots or violin-plots to provide
as much transparency as possible as to how data appear. This action is in
concordance with Martin C. Michel et. al. 2020,
*New Author Guidelines for Displaying Data and Reporting Data Analysis and Statistical Methods in Experimental Biology*.
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
