% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tSeries.R
\name{plot_tSeries}
\alias{plot_tSeries}
\title{Plot Time-Series Data for a Test Chemical and Corresponding Vehicle Control}
\usage{
plot_tSeries(
  pmr0,
  chemical,
  prsp = "SA",
  no.A = 10,
  unit.t = "2 min",
  unit.mov = "cm",
  unit.conc = paste0("μ", "M")
)
}
\arguments{
\item{pmr0}{- is a pmr0 dataset formatted as below
\itemize{
  \item srcf - name of file holding raw data
  \item acid - assay component id (Here ZFlmrALD-20-40-40)
    Zebrafish locomotor resonse, Acclimation/Light/Dark- _l_ minutes- _m_ minutes- _n_ minutes
  \item cpid - chemical name
  \item apid - assay plate id DNT###
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
   t = test, v = vehicle control
  \item wllq - well quality indicates if observation is viable for analysis
    1 = yes, 0 = no
  \item conc - concentration of chemical
  \item vtj - measurements at time period j for j /in {1,2,...,n}
    Here n = 50
}}

\item{chemical}{- string corresponding to name of chemical of interest}

\item{prsp}{- specifies SS or SA perspective, can take value "SS" or "SA"}

\item{no.A}{- number of time periods in acclimation period}

\item{unit.t}{- length of time contained in each time period. Defaults to "2 Minutes"}

\item{unit.mov}{- unit of distance traveled by subjects. Defaults to cm / 2 min}

\item{unit.conc}{- unit of chemical concentration. Defaults to $\mu$M}
}
\value{
A ggplot2 object displaying total movement time-series data from one of two perspectives
  SS
  \itemize {
    \item time series by individual
    \item time series of concentration-group means for each time period
    \item time series faceted by chemical concentration grouping
  }
  SA
  \itemize {
    \item time series of concentration-group means for each time period
    \item curves for each concentration are plotted together and are colored
    \item colored ribbons representing 50% confidence intervals about time period means
  }
}
\description{
Plots total movement per 2-minute bin time-series data of a test chemical and
it's corresponding vehicle control using either a Subject-Specific (SS)
perspective or a Sample Averaged (SA) perspective. User provides pmr0
formatted table, a character object specifying chemical to be plotted, and a
string object "SS" or "SA" specifying graphical method.
}
\details{
Created: 10/06/2021
Last edit: 08/18/2022

Subject Specific (SS) perspective fits curves to each individual's speed data
and plots. Population Averaged (SA) perspective fits to the mean speed at each
time period for a concentration group.
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
