% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplfit2_coreZR.R
\name{tcplfit2_coreZR}
\alias{tcplfit2_coreZR}
\title{Concentration-response curve fitting EDIT by Zachary Rowson}
\usage{
tcplfit2_coreZR(
  conc,
  resp,
  rmds,
  cutoff,
  force.fit = FALSE,
  bidirectional = TRUE,
  verbose = FALSE,
  do.plot = FALSE,
  fitmodels = c("cnst", "hill", "gnls", "poly1", "poly2", "pow", "exp2", "exp3",
    "exp4", "exp5"),
  ...
)
}
\arguments{
\item{conc}{Vector of concentrations (NOT in log units).}

\item{resp}{Vector of responses.}

\item{cutoff}{Desired cutoff. If no absolute responses > cutoff and
force.fit = F, will only fit constant model.}

\item{force.fit}{If force.fit = T, will fit all models regardless of cutoff.}

\item{bidirectional}{If bidirectional = F, will only give positive fits.}

\item{verbose}{If verbose = T, will print optimization details and aics.}

\item{do.plot}{If do.plot = T, will generate a plot comparing model curves.}

\item{fitmodels}{Vector of model names to try fitting. Missing models still
return a skeleton output filled with NAs.}

\item{...}{Other fitting parameters (deprecated).}

\item{name}{String of chemical name. #EDIT}

\item{assay}{String of assay name. #EDIT}
}
\value{
can.plot - logical indicating if plotting is possible

List of N(models) elements, one for each of the models run (up to 10),
followed by a last element "modelnames", which is a  vector of model names so
other functions can easily cycle through the output. For a full list, see the
documentation for the individual fitting method functions. For each model there
is a sublist with elements including:
  \itemize{
    \item success - was the model successfully fit
    \item aic - the AIC value
    \item cov - success of the the covariance matrix calculation
    \item rme - root mean error of the data around the curve
    \item modl - vector of model values at the civen concentrations
    \item tp - the top of the curve fit
    \item ga - the AC50 or Hill paramters
    \item er - the error term
    \item ... other paramters specific to the model (see the documentation for the specific models)
    \item tp_sd, ga_sd, p_sd, etc., the values of the standard deviations of the parameters for the models
    \item er_sd - standard deviation of the error term
    \item pars - the names of the parameters
    \item sds - the names of the standard deviations of the parameters
  }
}
\description{
Edit of original tcplfit2 function tcplfit2::tcplfit2_core. Editted to
remove default plotting method provided by tcplfit2. Replaces plotting method
with declaration of a logical to be passed to gabi::concRespCoreZR
wrapper function to determine if gabi::tcplggplotter should be run.
}
\details{
Last Edit: 08/18/2022
EDIT Tracking
Edits
1) add name, and assay to parameters
2) add can.plot to return
}
\examples{
conc <- c(.03, .1, .3, 1, 3, 10, 30, 100)
resp <- c(0, .1, 0, .2, .6, .9, 1.1, 1)
output <- tcplfit2_core(conc, resp, .8,
  fitmodels = c("cnst", "hill"), verbose = TRUE,
  do.plot = TRUE
)
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
