% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplfit2_funcfitZR.R
\name{tcplfit2_funcfitZR}
\alias{tcplfit2_funcfitZR}
\title{Produce curve fit function for tcpl fits}
\usage{
tcplfit2_funcfitZR(model, fit)
}
\arguments{
\item{model}{name of model to be transformed into function form.}

\item{fit}{params of model provided by fitting function, fit___.
Entries in params depend upon the model fit.}
}
\value{
func - function form of model params provided
}
\description{
Returns a function with fitted parameters from tcplfit2 model fitting.
Goal here is to provide a function for visualization with
gabi::tcplggplotter
}
\details{
Created: 09/09/2021
Last edit: 09/09/2021
Roxygen created this manual page on `r Sys.Date()` using R version
`r getRversion()`.
All models are equal to 0 at 0 concentration (zero background).
}
\examples{
conc <- c(.03, .1, .3, 1, 3, 10, 30, 100)
resp <- c(0, .1, 0, .2, .6, .9, 1.1, 1)
output <- tcplfit2_core(conc, resp, .8,
  fitmodels = c("cnst", "hill"), verbose = TRUE,
  do.plot = TRUE
)
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
