% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplggplotter.R
\name{tcplggplotter}
\alias{tcplggplotter}
\title{Plot tcplfit2 output with ggplot2}
\usage{
tcplggplotter(
  endp,
  resp,
  bresp,
  conc,
  logc,
  rmds,
  bmed,
  lam.hat,
  shift,
  params,
  summary,
  verbose.plot,
  unit.conc = paste0("μ", "M")
)
}
\arguments{
\item{resp}{Vector of test-group response extracted from tcpl row object}

\item{bresp}{Vector of baseline response extracted from tcpl row object}

\item{logc}{Log base 10 of vector of concentrations from tcpl row object}

\item{rmds}{Response medians of each test-concentration group}

\item{bmed}{Response median of baseline}

\item{lam.hat}{Numeric power used to Box-Cox transform data}

\item{shift}{Numeric value used to shift data values from zero for transformation}

\item{params}{List of model-fit parameters returned from tcplfit2::tcplfit2_core.
List of N(models) elements, one for each of the models run (up to 10),
followed by a last element "modelnames", which is a  vector of model names so
other functions can easily cycle through the output. For a full list, see the
documentation for the individual fitting method functions. For each model there
is a sublist with elements including:
  \itemize{
    \item success - was the model successfully fit
    \item aic - the AIC value
    \item cov - success of the the covariance matrix calculation
    \item rme - root mean error of the data around the curve
    \item modl - vector of model values at the given concentrations
    \item tp - the top of the curve fit
    \item ga - the AC50 or Hill paramters
    \item er - the error term
    \item ... other paramters specific to the model (see the documentation for the specific models)
    \item tp_sd, ga_sd, p_sd, etc., the values of the standard deviations of the paramters for the models
    \item er_sd - standard deviation of the error term
    \item pars - the names of the parameters
    \item sds - the names of the standard deviations of the paramters
  }}

\item{summary}{list of tcplfit2 analysis output returned from tcplfit2::tcplhit2_core}

\item{verbose.plot}{If TRUE, ggplots will have a caption listing summarizing statistics (default TRUE)}

\item{unit.conc}{unit of chemical concentration (default $\mu$M)}
}
\value{
A ggplot2 object of tcplfit2 analysis displaying...
  \itemize {
    \item model-fits distinguished by color
    \item cutoff range respresented by gray box
    \item response by concentration
    \item median response by concentration group and \eqn{90 \leq} CI's
    \item hitcall produced by tcplfit2
    \item best fit by AIC
    }
}
\description{
Plots concentration response trend of multiple-concentration experiment
on a chemical. tcplfit2 functions are used to produce interpolating
fits on chemical response.
  \itemize {
  \item cnst - constant model
  \item hill - hill
  \item gnls - gain-loss
  \item poly1 - polynomial of 1st degree
  \item poly2 - polynomial of 2nd degree
  \item pwr - power function
  \item exp2 - exponential of 2nd degree
  \item exp3 - exponential of 3rd degree
  \item exp4 - exponential of 4th degree
  \item exp5 - exponential of 4th degree
  }
}
\details{
Last edit: 12/13/2021
Roxygen created this manual page on `r Sys.Date()` using R version
`r getRversion()`.
Used in conjunction with concRespCoreZR() and tcplfit2_coreZR().
Not meant to be used as a stand alone function. Graphics are meant to replicate
appearance of default graphics produced in tcplfit2::tcplfit2_core.
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
