% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_Bhv(copy).R, R/upload_Bhv.R
\name{upload_Bhv}
\alias{upload_Bhv}
\title{COPY
This function uploads and formats behavioral data. Behavioral data are stored
in one .xlsx file names "All Chemicals one per sheet.xlsx". This .xlsx contains
unique worksheets with each sheet containing behavioral data of one test
chemical and it's corresponding control. These worksheets are imported as
data.tables and are named by the test chemical they describe.}
\usage{
upload_Bhv(path)

upload_Bhv(path)
}
\arguments{
\item{path}{this is the path to "All Chemicals one per sheet.xlsx"}
}
\description{
COPY
This function uploads and formats behavioral data. Behavioral data are stored
in one .xlsx file names "All Chemicals one per sheet.xlsx". This .xlsx contains
unique worksheets with each sheet containing behavioral data of one test
chemical and it's corresponding control. These worksheets are imported as
data.tables and are named by the test chemical they describe.

This function uploads and formats behavioral data. Behavioral data are stored
in one .xlsx file names "All Chemicals one per sheet.xlsx". This .xlsx contains
unique worksheets with each sheet containing behavioral data of one test
chemical and it's corresponding control. These worksheets are imported as
data.tables and are named by the test chemical they describe.
}
