% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_DNT60.R
\name{upload_DNT60}
\alias{upload_DNT60}
\title{Upload Padilla DNT60 Behavior Data from "All Chemicals on per sheet.xlsx"}
\usage{
upload_DNT60(path)
}
\arguments{
\item{path}{path to Excel file "All Chemicals one per sheet.xlsx"}
}
\value{
A pmr0 format data.table of all test animals described in "All Chemicals one per sheet.xlsx".
  Returned table will contain the following variables.
  \itemize{
    \item srcf - name of file that is being formatted
    \item acid - assay component id (Here ZFpmrALD-20-40-40)
      Zebrafish photomotor resonse, AccilmationLightDark-20 minutes-40 minutes-40 minutes
    \item cpid - chemical name
    \item apid - assay plate id, DNT###
    \item rowi - row on plate
    \item coli - column on plate
    \item wllt - well type according to tcpl mc0 format
     t = test, v = vehicle control
    \item wllq - well quality indicates if observation is viable for analysis
      1 = yes, 0 = no
    \item conc - concentration of chemical
    \item tj - measurements of total distance moved at 2 minute time period j for j /in {1,2,...,n}
      Here n = 50
  }

  @import data.table
}
\description{
Upload Padilla DNT60 Behavior Data from "All Chemicals on per sheet.xlsx"
}
\details{
Last edit: 05/06/2021
Roxygen created this manual page on `r Sys.Date()` using R version
`r getRversion()`.
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}

Returns complete USA EPA ORD CCTE BCTD RADB Padilla lab DNT60 behavior data
set in pmr0 format. pmr0 format is easily transformed
into mc0 with DNT.60.Analysis::as_mc0() for tcpl analysis.
}
