% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatis_Endp.R
\name{whatis_Endp}
\alias{whatis_Endp}
\title{Expand endpoint acronyms for clarity}
\usage{
whatis_Endp(endp)
}
\arguments{
\item{endp}{string of endpoint name in acronym form}
}
\value{
endpoint name in expanded form
}
\description{
Expands endpoint acronym when given acronym in string form.
}
\details{
Created 11/02/2021. Last edit 12/13/21.
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
