% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_bxcx.R
\name{apply_bxcx}
\alias{apply_bxcx}
\title{Apply Box-Cox transformation to rval}
\usage{
apply_bxcx(data)
}
\arguments{
\item{data}{is a mc0 dataset formatted as below.
\itemize{
  \item srcf - name of file that is being formatted
  \item acid - assay component id (Here ZFlmrA/L/D-20-40-40)
    Zebrafish locomotor response, Acclimation/Light/Dark- _l_ minutes- _m_ minutes- _n_ minutes
  \item cpid - chemical name
  \item apid - assay plate id DNT###
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
   t = test, v = vehicle control, c/o = positive control
  \item wllq - well quality indicates if observation is viable for analysis
    1 = yes, 0 = no
  \item conc - concentration of chemical
  \item rval - response value, endpoint of interest
}}
}
\value{
mc0 formatted table with Box-Cox transformed rvals
}
\description{
Applies Box-Cox power transformation to mc0 rval with goal
of normalizing data. Requires estimating an optimal
lambda parameter, lambda.hat. Lambda.hat is chosen via
MLE using control data. Shift parameter is
included if data is less than or equal to 0. Shift
parameter equals floor(minimum observed rval).
}
\details{
Last edit: 08/18/2022
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
