% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mc0.R
\name{as_mc0}
\alias{as_mc0}
\title{Format pmr0 data set as mc0 for tcpl analysis}
\usage{
as_mc0(table, rval = NULL, no.A = 10, no.L = 20, no.D = 20)
}
\arguments{
\item{table}{is a pmr0 dataset formatted as below
\itemize{
  \item srcf - name of file that is being formatted
  \item acid - assay component id (ZFpmrALD-l-m-n)
    Zebrafish photomotor resonse, Accilmation/Light/Dark-l minutes-m minutes-n minutes
  \item cpid - chemical name
  \item apid - assay plate id
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
   t = test, v = vehicle control, c/o = positive control
  \item wllq - well quality indicates if observation is viable for analysis
    1 = yes, 0 = no
  \item conc - concentration of chemical
  \item vtj - measurements at time period j for j /in {1,2,...,n}
    Here n = 50
}}

\item{rval}{is the name of desired endpoint as a string}

\item{no.A}{number of measurements made in acclimation period, defaults to 10}

\item{no.L}{number of measurements made in light period, defaults to 20}

\item{no.D}{number of measurements made in dark period, defaults to 20}
}
\value{
A mc0 format data.table
  \itemize{
    \item srcf - name of file that is being formatted
    \item acid - assay component id (ZFpmrA/L/D-l-m-n)
      Zebrafish photomotor resonse, Accilmation/Light/Dark-l minutes-m minutes-n minutes
    \item cpid - chemical name
    \item apid - assay plate id
    \item rowi - row on plate
    \item coli - column on plate
    \item wllt - well type according to tcpl mc0 format
     t = test, v = vehicle control
    \item wllq - well quality indicates if observation is viable for analysis
      1 = yes, 0 = no, c/o = positive control
    \item conc - concentration of chemical
    \item rval - endpoints resp values of each fish
  }
}
\description{
Formats a pmr0 formatted dataset into mc0 format for
tcpl analysis. Requires the creation of endpoints via
gabi::create_endp().
}
\details{
Last Edit: 02/11/2022
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
