% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_row.R
\name{as_row}
\alias{as_row}
\title{Format chemical data into tcpl row object for tcpl analysis}
\usage{
as_row(data, chemical, endp, lam.hat = 1, shift = 0)
}
\arguments{
\item{data}{is a mc0 dataset formatted as below
\itemize{
  \item srcf - name of file that is being formatted
  \item acid - assay component id (Here ZFlmrALD-20-40-40)
    Zebrafish locomotor response, Acclimation/Light/Dark- _l_ minutes- _m_ minutes- _n_ minutes
  \item cpid - chemical name
  \item apid - assay plate id DNT###
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
   t = test, v = vehicle control, c/o = positive control
  \item wllq - well quality indicates if observation is viable for analysis
    1 = yes, 0 = no
  \item conc - concentration of chemical
  \item rval - endpoints resp values of each fish
}}

\item{chemical}{is a string representing chemical of interest.}

\item{endp}{is a string representing the endpoint of interest.}

\item{lam.hat}{is power used to Box-Cox transform data.}

\item{shift}{is value used to shift data values from zero for transformation.}
}
\value{
A tcpl row object
  \itemize{
      \item conc - vector of concentrations at which each fish was tested
      \item resp - response of each fish ordered to align with conc vector
      \item bresp - baseline response: vector of baseline fish response values
      \item bmed - median of baseline response
      \item cutoff - numerical value representing noise about baseline
      \item onesd - standard deviation of bresp
      \item name - name of tested chemical
      \item assay - name of assay and endpoint
  }

  @import data.table
}
\description{
Formats a chemical data from a mc0 dataset into a row
object for tcpl analysis. Row objects are lists of
data and descriptive statistics necessary for tcpl
analysis.
}
\details{
Created 06/01/2021 Last edit: 08/18/2022
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
