% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hbt.R
\name{calc_hbt}
\alias{calc_hbt}
\title{Calculate measures of habituation per light level for individual fish}
\usage{
calc_hbt(data, no.A = 10, no.L = 20, no.D = 20)
}
\arguments{
\item{data}{is a pmr0 formatted data.table
\itemize{
  \item srcf - name of file that is being formatted
  \item acid - assay component id
  \item cpid - chemical name
  \item apid - assay plate id
  \item rowi - row on plate
  \item coli - column on plate
  \item wllt - well type according to tcpl mc0 format
  \item wllq - well quality indicates if observation is viable for analysis
  \item conc - concentration of chemical
  \item tj - measurements at time period j for j /in {1,2,3,...,n}
}}

\item{no.A}{number of measurements made in acclimation period}

\item{no.L}{number of measurements made in light period}

\item{no.D}{number of measurements made in dark period}
}
\value{
A list with habituation ednpoints for each fish
  \itemize{
    \item hbt_L - speed ratio in light. max(speed) / min(speed)
    \item hbt_D - speed ratio in dark. max(speed) / min(speed)
  }
}
\description{
Calculates measures of habituation per light period (light or dark).
User provides a lmr0 table and inputs
number of time periods in each experimental stage:
acclimation, light, and dark.
}
\details{
hbt_L = max(speed in light) / min(speed in light), hbt_D = max(speed in dark) / min(speed in dark).
Created 09/28/2021. Last edit 08/18/2022.
}
\author{
Zachary Rowson \email{Rowson.Zachary@epa.gov}
}
